+++ src/checkbox/checkbox.ts
@@ -14,6 +14,8 @@ export default class CheckBox extends SuperComponent {
     `${prefix}-class-border`,
   ];
 
+  behaviors = ['wx://form-field'];
+
   relations = {
     '../checkbox-group/checkbox-group': {
       type: 'ancestor' as 'ancestor',
@@ -35,7 +37,7 @@ export default class CheckBox extends SuperComponent {
   // 组件的内部数据
   data = {
     classPrefix,
-    classBasePrefix: prefix,
+    prefix,
     active: false,
     halfChecked: false,
     optionLinked: false,