+++ src/image/index.ts
@@ -1,0 +1,5 @@
+import { TdImageProps } from './type';
+
+export type ImageProps = TdImageProps;
+export * from './props';
+export * from './image';