+++ src/radio/radio.ts
@@ -19,6 +19,8 @@ export default class Radio extends SuperComponent {
     `${prefix}-class-border`,
   ];
 
+  behaviors = ['wx://form-field'];
+
   relations = {
     '../radio-group/radio-group': {
       type: 'ancestor' as 'ancestor',
@@ -58,7 +60,6 @@ export default class Radio extends SuperComponent {
     prefix,
     active: false,
     classPrefix: name,
-    classBasePrefix: prefix,
     customIcon: false,
     optionLinked: false,
     iconVal: [],