+++ tdesign-miniprogram/message/message.interface.js
@@ -1,9 +1,0 @@
-export var MessageType;
-(function (MessageType) {
-    MessageType["info"] = "info";
-    MessageType["success"] = "success";
-    MessageType["warning"] = "warning";
-    MessageType["error"] = "error";
-})(MessageType || (MessageType = {}));
-
-//# sourceMappingURL=message.interface.js.map