+++ tdesign-miniprogram/swiper/common/constants.js
@@ -1,9 +1,0 @@
-const DIRECTION = {
-    // 水平
-    HOR: 'horizontal',
-    // 垂直
-    VER: 'vertical',
-};
-export { DIRECTION };
-
-//# sourceMappingURL=constants.js.map